Clean up: one slow line, two oversized modules, one duplicated helper
A performance and readability pass over the whole project. No behaviour
changes: the composite still reports 83.4 km/h against the quickest
pilot's 3:38:43, the same as before.

**The slow line.** Profiling a twenty-pilot day put 10 of 18 seconds in
one place: aerofiles' TimeZoneFix.utcoffset, called 16 million times. It
builds a fresh timedelta on every call, and every comparison between two
timezone-aware datetimes calls it. Re-stamping each fix with the standard
library's UTC at load fixes it outright, and aerofiles says in the class
itself that it exists for Python 2.6 and can be replaced once 3.x is the
minimum — which it is here. This was the whole ballgame: pooling the
field's climbs went from 4.4 s to 0.1 s on that change alone.

**The quadratic scans.** "What happened between these two moments" was
asked once per leg, once per thermal and once per stretch, and each time
it scanned the whole trace. core/trace.py answers it by binary search
instead, and now says so once in a named place rather than five times as
a comprehension. Pooling the field's climbs went on to 0.01 s, its routes
to 0.04 s.

Together, on twenty flights of 10,160 fixes: load and analyse 6.2 -> 3.5 s,
the gap to a reference 2.2 -> 0.4 s, the field's climbs 4.4 -> 0.01 s, its
routes 4.2 -> 0.04 s, the composite 1.9 -> 0.2 s. The test suite went from
65 s to 40 s as a side effect. What remains is inside aerofiles and
opensoar.

**The oversized modules.** maps.py (782 lines) is now a package of five:
base, layers, flight, gap, field. day_view.py (639) is a package of five:
archive, tables, gap_tab, field_tab, view. Both keep their public surface
exactly as it was, so nothing that imports them changed. Helpers that
cross a module boundary lost their leading underscore, because a private
name shared between five files is not private.

**The duplicated helper.** Two modules carried a _duration that formatted
differently and disagreed about what to do with zero, so the same quantity
read one way on one tab and another on the next. app/format.py now holds
number/gap/clock, and the rule they share is that missing and zero are
different: a metric nobody could compute shows an em dash, one that came
out zero shows zero. Collapsing the two is how "no circling at all" came
to look like "we could not tell". Fixed a negative-zero gap while I was
there — half a second behind printed as "-0:00".

**Dead code and lint.** Removed six accessors and one constant nothing
read, and replaced Altitude.is_ground (unused) with is_agl, which the
airspace band check was open-coding twice. Widened the ruff rule set from
five families to thirteen and fixed the 85 findings; the interesting one
was a shared mutable default on the fake SoaringSpot server, where one
list was serving every instance's request log.

New tests for both new modules: 18 for the trace helpers, held against the
comprehensions they replaced as an oracle, and 12 for the formatters.
289 tests pass.

Verified in the browser after the split: both views, all five tabs, no
page errors, and the same numbers as before.

+++ tests/test_replay.py
@@ -55,8 +55,8 @@ def test_thinning_respects_the_target_and_keeps_the_ends(flights):
     for index, track in enumerate(small["tracks"]):
         assert len(track["t"]) <= 260, "thinning overshot its target"
         trace = flights[index].flight.trace
-        assert track["t"][0] == int(round(trace[0]["datetime"].timestamp()))
-        assert track["t"][-1] == int(round(trace[-1]["datetime"].timestamp()))
+        assert track["t"][0] == round(trace[0]["datetime"].timestamp())
+        assert track["t"][-1] == round(trace[-1]["datetime"].timestamp())
 
 
 def test_clock_alignment_puts_pilots_on_one_timeline(clock_payload, flights):
@@ -74,7 +74,7 @@ def test_start_alignment_measures_from_each_pilots_own_start(flights):
     for index, track in enumerate(aligned["tracks"]):
         start = flights[index].start_time
         trace = flights[index].flight.trace
-        expected_first = int(round((trace[0]["datetime"] - start).total_seconds()))
+        expected_first = round((trace[0]["datetime"] - start).total_seconds())
         assert track["t"][0] == expected_first
         # Pre-start time is negative; the task itself begins at zero.
         assert track["t"][0] < 0 < track["t"][-1]
